extract translation-keys-tree-view
